+++ cli/srecli/group/users.py
@@ -52,7 +52,7 @@ def create(ctx: ContextObj, username, password, role_name, customer_id):
 
 @users.command(cls=ViewCommand, name='update')
 @click.option('--username', required=True, type=str,
-              help='Username to create user')
+              help='Username to update user')
 @click.option('--password', '-p', type=str, help='New user password')
 @click.option('--role_name', '-rn', type=str,
               help='Role to assign to this user. '
@@ -72,7 +72,7 @@ def update(ctx: ContextObj, username, customer_id, password, role_name):
 
 @users.command(cls=ViewCommand, name='delete')
 @click.option('--username', required=True, type=str,
-              help='Username to create user')
+              help='Username to delete user')
 @cli_response()
 def delete(ctx: ContextObj, username, customer_id):
     """